Integrated ater: kvcache pa gemm rmsnorm

+++ vllm/model_executor/models/mixtral.py
@@ -98,7 +98,7 @@ class MixtralMoE(nn.Module):
         orig_shape = hidden_states.shape
         hidden_states = hidden_states.view(-1, self.hidden_size)
         # router_logits: (num_tokens, n_experts)
-        router_logits, _ = self.gate(hidden_states)
+        router_logits, _ = self.gate(hidden_states, out_dtype = torch.float32)
         final_hidden_states = self.experts(hidden_states, router_logits)
         return final_hidden_states.view(orig_shape)
